Game State Update
Added update for the gamestate to loop through properly but it does not
send to the client and back yet. This will need to be implemented. I
noted key points where that would need to be changed, everything else
should work as long the ending and receving is solid.

Also added a quick "ClientDesign" class that does nothing but show off
what I think the ClientDesign looping should look like. Let me know if
you need more clarification.